Move interfaces out of libcloud specific files.
