+++ flocker/provision/__init__.py
@@ -4,7 +4,7 @@
 Provisioning for acceptance tests.
 """
 
-from ._common import PackageSource, Variants
+from ._common import PackageSource, Variants, INode, IProvisioner
 from ._install import provision, configure_cluster
 from ._rackspace import rackspace_provisioner
 from ._aws import aws_provisioner
@@ -17,6 +17,7 @@ CLOUD_PROVIDERS = {
 
 __all__ = [
     'PackageSource', 'Variants',
+    'INode', 'IProvisioner',
     'provision',
     'CLOUD_PROVIDERS',
     'configure_cluster',
