+++ flocker/provision/_aws.py
@@ -9,8 +9,12 @@ from time import time, sleep
 
 from pyrsistent import PClass, field
 
+from zope.interface import implementer
+
 from effect.retry import retry
 from effect import Effect, Constant
+
+from ._common import INode, IProvisioner
 
 from ._install import (
     provision,
@@ -36,6 +40,7 @@ IMAGE_NAMES = {
 }
 
 
+@implementer(INode)
 class AWSNode(PClass):
     _provisioner = field(mandatory=True)
     _instance = field(mandatory=True)
@@ -124,6 +129,7 @@ class AWSNode(PClass):
         ).on(success=do_reboot)
 
 
+@implementer(IProvisioner)
 class AWSProvisioner(PClass):
     connection = field(mandatory=True)
     keyname = field(mandatory=True)
@@ -145,6 +151,8 @@ class AWSProvisioner(PClass):
     def create_node(self, name, distribution,
                     size=None, disk_size=8,
                     metadata={}):
+        # Import these here, so that this can be imported without installng
+        # libcloud.
         from boto.ec2.blockdevicemapping import (
             EBSBlockDeviceType, BlockDeviceMapping,
         )
@@ -223,6 +231,8 @@ def aws_provisioner(access_key, secret_access_token, keyname,
     :param list security_groups: List of security groups to put created nodes
         in.
     """
+    # Import these here, so that this can be imported without installng
+    # libcloud.
     from boto.ec2 import connect_to_region
     conn = connect_to_region(
         region,
